Add files via upload
This takes the 3d shapes dataset and attempts to encode/decode it using the ELBO loss function. This implementation is like that found on Tensorflows demo implementation but amends it for an image dataset with three colour channels. The reproduced images are not the most elegant due to the small number of images, as well as the modest size of the autoencoder network, as well as the latent dimension size of 2. Enhancements can be made my playing around with aforementioned elements.